requirements.txt
requests~=2.27
GitPython~=3.1